Fixed compilation in VS2015

+++ SoftwareRender/vec3.cpp
@@ -38,7 +38,7 @@ namespace math
 
 	float vec3::length() const
 	{
-		return sqrt(x*x + y*y + z*z);
+		return sqrtf(x*x + y*y + z*z);
 	}
 
 }
